Issue-3499: Move signals module to compose/common package
Partially implements #3499
