+++ compose/cli/main.py
@@ -12,7 +12,6 @@ from inspect import getdoc
 from operator import attrgetter
 
 from . import errors
-from . import signals
 from .. import __version__
 from ..common.const import DEFAULT_TIMEOUT
 from ..common.const import IS_WINDOWS_PLATFORM
@@ -42,6 +41,7 @@ from .log_printer import LogPrinter
 from .utils import get_version_info
 from .utils import yesno
 from compose.cli.progress_stream import StreamOutputError
+from compose.common import signals
 
 if not IS_WINDOWS_PLATFORM:
     from dockerpty.pty import PseudoTerminal, RunOperation, ExecOperation
