+++ compose/common/signals.py
